phase A: split .js
