+++ custom_components/ramses_extras/features/device_simulator/www/device_simulator/card/constants.js
@@ -1,0 +1,10 @@
+/**
+ * Device Simulator Card - shared constants.
+ */
+
+export const SCENARIO_MANUAL_DEVICE = "autonomous_emissions";
+export const SCENARIO_PROFILE_EMISSIONS = "profile_emissions";
+export const SCENARIO_LOAD_PROFILE = "load_profile_yaml";
+export const SCENARIO_DEVICE_PLAYBACK = "device_playback";
+export const LOADER_DRAFT_KEY = "ramsesExtras.deviceSimulator.loaderDraft";
+export const MESSAGE_HISTORY_LIMIT = 80;
