+++ custom_components/ramses_extras/features/device_simulator/www/device_simulator/device-simulator-card.js
@@ -4,128 +4,20 @@
  */
 
 import { RamsesBaseCard } from '../../helpers/ramses-base-card.js';
+import {
+  SCENARIO_MANUAL_DEVICE,
+  SCENARIO_PROFILE_EMISSIONS,
+  SCENARIO_LOAD_PROFILE,
+  SCENARIO_DEVICE_PLAYBACK,
+  LOADER_DRAFT_KEY,
+  MESSAGE_HISTORY_LIMIT,
+} from './card/constants.js';
+import { CARD_STYLE } from './card/styles.js';
+import { buildProfiles } from './card/tabs/profiles-tab.js';
+import { buildScenarios } from './card/tabs/scenarios-tab.js';
+import { buildDevices } from './card/tabs/devices-tab.js';
+import { buildEvents } from './card/tabs/events-tab.js';
 
-const SCENARIO_MANUAL_DEVICE = "autonomous_emissions";
-const SCENARIO_PROFILE_EMISSIONS = "profile_emissions";
-const SCENARIO_LOAD_PROFILE = "load_profile_yaml";
-const SCENARIO_DEVICE_PLAYBACK = "device_playback";
-const LOADER_DRAFT_KEY = "ramsesExtras.deviceSimulator.loaderDraft";
-const MESSAGE_HISTORY_LIMIT = 80;
-
-const CARD_STYLE = `
-  :host { display: block; padding: 16px; }
-  .header { display: flex; align-items: center; justify-content: space-between; margin-bottom: 16px; }
-  .title { font-size: 1.2em; font-weight: 500; }
-  .badge { padding: 4px 12px; border-radius: 12px; font-size: 0.75em; text-transform: uppercase; }
-  .badge.active { background: var(--success-color, #4caf50); color: white; }
-  .badge.running { background: var(--primary-color, #03a9f4); color: white; animation: pulse 1.5s infinite; }
-  @keyframes pulse { 0%, 100% { opacity: 1; } 50% { opacity: 0.6; } }
-  .tabs { display: flex; gap: 8px; margin-bottom: 16px; border-bottom: 1px solid var(--divider-color); }
-  .tab { padding: 12px 16px; cursor: pointer; border-bottom: 2px solid transparent; font-size: 0.9em; }
-  .tab.active { color: var(--primary-color); border-bottom-color: var(--primary-color); }
-  .content { display: none; }
-  .content.active { display: block; }
-  .btn { padding: 8px 16px; border: none; border-radius: 4px; cursor: pointer; font-size: 0.85em; }
-  .btn-primary { background: var(--primary-color); color: var(--text-primary-color); }
-  .btn-secondary { background: var(--secondary-background-color); color: var(--primary-text-color); border: 1px solid var(--divider-color); }
-  .btn-danger { background: var(--error-color); color: white; }
-  .grid { display: grid; gap: 12px; }
-  .devices-grid { grid-template-columns: repeat(auto-fit, minmax(720px, 1fr)); }
-  .card { border: 1px solid var(--divider-color); border-radius: 8px; padding: 16px; }
-  .card:hover { border-color: var(--primary-color); }
-  .card.active { border-color: var(--success-color); background: var(--success-color-light); }
-  .stats { display: grid; grid-template-columns: repeat(4, 1fr); gap: 12px; margin-bottom: 16px; }
-  .stat { background: var(--secondary-background-color); padding: 16px; border-radius: 8px; text-align: center; }
-  .stat-value { font-size: 1.5em; font-weight: 600; color: var(--primary-color); }
-  .event-log { max-height: 300px; overflow-y: auto; border: 1px solid var(--divider-color); border-radius: 8px; }
-  .event { display: flex; gap: 12px; padding: 8px; border-bottom: 1px solid var(--divider-color); font-size: 0.85em; }
-  .event.unavailable { background: var(--error-color-light); border-left: 3px solid var(--error-color); }
-  .event-type { padding: 2px 6px; border-radius: 4px; font-size: 0.75em; text-transform: uppercase; }
-  .event-type.rx { background: var(--success-color-light); color: var(--success-color); }
-  .event-type.tx { background: var(--info-color-light); color: var(--info-color); }
-  .event-type.unavailable { background: var(--error-color-light); color: var(--error-color); }
-  .device-actions { display: flex; gap: 8px; margin-top: 8px; align-items: center; flex-wrap: wrap; }
-  .device-emitter { display: flex; gap: 8px; align-items: center; flex-wrap: wrap; margin-top: 6px; font-size: 0.8em; }
-  .toggle { display: flex; align-items: center; gap: 6px; cursor: pointer; font-size: 0.85em; }
-  .codes-row { display: flex; gap: 6px; flex-wrap: wrap; margin-top: 6px; align-items: center; }
-  .code-chip { display: flex; align-items: center; gap: 4px; padding: 2px 8px; border-radius: 12px; font-size: 0.75em; background: var(--secondary-background-color); border: 1px solid var(--divider-color); }
-  .code-chip button { background: none; border: none; cursor: pointer; padding: 0; font-size: 0.9em; color: var(--secondary-text-color); line-height: 1; }
-  .add-code { display: flex; gap: 4px; margin-top: 4px; }
-  .add-code input { padding: 3px 6px; border: 1px solid var(--divider-color); border-radius: 4px; font-size: 0.8em; width: 80px; background: var(--card-background-color); color: var(--primary-text-color); }
-  .device-controls { margin-bottom: 16px; }
-  .device-list-empty { color: var(--secondary-text-color); padding: 8px 0; }
-  .profile-missing { margin-top: 12px; padding: 8px 12px; border-radius: 8px; background: var(--secondary-background-color); display: flex; flex-direction: column; gap: 8px; }
-  .profile-missing strong { font-size: 0.85em; color: var(--secondary-text-color); }
-  .profile-missing-entry { display: flex; gap: 8px; align-items: center; justify-content: space-between; flex-wrap: wrap; font-size: 0.8em; }
-  .speed-card { display:flex; flex-direction:column; gap:10px; }
-  .speed-card .speed-controls { display:flex; flex-wrap:wrap; gap:8px; align-items:center; }
-  .speed-card input[type='range'] { width:200px; }
-  .speed-card input[type='number'] { width:90px; padding:4px; border:1px solid var(--divider-color); border-radius:4px; background:var(--card-background-color); color:var(--primary-text-color); }
-  .speed-card button { font-size:0.8em; }
-  .chip { display: inline-flex; align-items: center; padding: 2px 8px; border-radius: 999px; font-size: 0.7em; font-weight: 600; text-transform: uppercase; }
-  .chip.profile { background: var(--primary-color); color: white; }
-  .chip.manual { background: var(--info-color, #0288d1); color: white; }
-  .chip.known { background: var(--success-color, #388e3c); color: white; }
-  .chip.sensor { background: var(--warning-color, #ffa000); color: #000; }
-  .chip.reply { background: var(--info-color, #0288d1); color: #fff; }
-  .chip.disabled { background: var(--divider-color); color: var(--secondary-text-color); }
-  .chip.muted { background: var(--divider-color); color: var(--secondary-text-color); }
-  .chip.source { background: var(--secondary-background-color); color: var(--secondary-text-color); }
-  .chip.emitting { background: var(--success-color, #388e3c); color: #fff; }
-  .chip.idle { background: var(--warning-color, #ffa000); color: #000; }
-  .chip.silenced { background: var(--error-color, #f44336); color: #fff; }
-  .btn[disabled] { opacity: 0.6; cursor: not-allowed; }
-  .scenario-form { margin-top: 8px; display: flex; flex-direction: column; gap: 8px; }
-  .scenario-field { display: flex; flex-direction: column; gap: 4px; font-size: 0.8em; }
-  .scenario-field label { color: var(--secondary-text-color); font-weight: 600; display: flex; align-items: center; justify-content: space-between; gap: 8px; }
-  .scenario-field input[type="text"],
-  .scenario-field input[type="search"],
-  .scenario-field input[type="number"] { padding: 4px 8px; border: 1px solid var(--divider-color); border-radius: 4px; background: var(--card-background-color); color: var(--primary-text-color); font-size: 0.9em; }
-  .scenario-field select { padding: 4px 8px; border: 1px solid var(--divider-color); border-radius: 4px; background: var(--card-background-color); color: var(--primary-text-color); font-size: 0.9em; }
-  .scenario-field textarea { padding: 8px; border: 1px solid var(--divider-color); border-radius: 6px; min-height: 140px; font-family: var(--code-font-family, 'Fira Code', monospace); font-size: 0.85em; background: var(--card-background-color); color: var(--primary-text-color); resize: vertical; line-height: 1.35; }
-  .scenario-field--checkbox { flex-direction: row; align-items: center; gap: 8px; }
-  .scenario-field--checkbox label { font-weight: 500; margin: 0; }
-  .scenario-description { margin-top: 6px; font-size: 0.8em; color: var(--secondary-text-color); }
-  .msg-log {
-    font-family: var(--code-font-family, 'Fira Code', monospace);
-    font-size: 0.75em;
-    max-height: 260px;
-    overflow: auto;
-    border: 1px solid var(--divider-color);
-    border-radius: 6px;
-    width: 100%;
-    max-width: 100%;
-  }
-  .msg-log-table { width: max-content; min-width: 100%; border-collapse: collapse; }
-  .msg-log-table th,
-  .msg-log-table td { padding: 4px 8px; white-space: nowrap; border-bottom: 1px solid var(--divider-color); text-align: left; user-select: text; -webkit-user-select: text; }
-  .msg-log-table thead th { font-weight: 600; color: var(--secondary-text-color); font-size: 0.78em; background: var(--secondary-background-color); position: sticky; top: 0; z-index: 1; }
-  .msg-log-table tbody tr:last-child td { border-bottom: none; }
-  .msg-log-table tbody tr.in td { background: rgba(var(--rgb-success-color, 76,175,80), 0.07); }
-  .msg-log-table tbody tr.out td { background: rgba(var(--rgb-info-color, 3,169,244), 0.07); }
-  /* Origin-based row tinting (overrides direction tint where applicable). */
-  .msg-log-table tbody tr.origin-rf         td { background: rgba(76, 175, 80, 0.10); }
-  .msg-log-table tbody tr.origin-sim        td { background: rgba(3, 169, 244, 0.10); }
-  .msg-log-table tbody tr.origin-auto_answer td { background: rgba(255, 152, 0, 0.14); }
-  .msg-log-table tbody tr.origin-auto_emit  td { background: rgba(156, 39, 176, 0.12); }
-  .msg-origin { font-weight: 600; font-size: 0.72em; text-transform: uppercase; padding: 1px 6px; border-radius: 8px; letter-spacing: 0.02em; white-space: nowrap; }
-  .msg-origin.origin-rf         { background: rgba(76, 175, 80, 0.20); color: #2e7d32; }
-  .msg-origin.origin-sim        { background: rgba(3, 169, 244, 0.20); color: #01579b; }
-  .msg-origin.origin-auto_answer { background: rgba(255, 152, 0, 0.25); color: #e65100; }
-  .msg-origin.origin-auto_emit  { background: rgba(156, 39, 176, 0.22); color: #6a1b9a; }
-  .msg-dir { font-weight: 700; font-size: 0.8em; text-transform: uppercase; }
-  .msg-dir.in { color: var(--success-color, #4caf50); }
-  .msg-dir.out { color: var(--info-color, #0288d1); }
-  .msg-verb { font-weight: 600; color: var(--primary-color); }
-  .msg-code { font-weight: 600; }
-  .msg-addr { color: var(--secondary-text-color); font-size: 0.9em; overflow: hidden; text-overflow: ellipsis; max-width: 200px; }
-  .msg-payload { color: var(--secondary-text-color); font-size: 0.9em; overflow: hidden; text-overflow: ellipsis; max-width: 300px; }
-  .msg-ts { color: var(--disabled-color, #9e9e9e); font-size: 0.85em; }
-  .msg-log-empty { padding: 12px; color: var(--secondary-text-color); font-size: 0.85em; }
-  .msg-log-header { display: flex; align-items: center; justify-content: space-between; gap: 8px; margin-bottom: 6px; }
-  .msg-log-header-row { display:flex; align-items:center; justify-content: space-between; gap:8px; margin-bottom:6px; }
-  .msg-log-header-row button { font-size:0.7em; padding:4px 8px; }
-`;
 
 class DeviceSimulatorCard extends RamsesBaseCard {
   constructor() {
@@ -188,44 +80,6 @@ class DeviceSimulatorCard extends RamsesBaseCard {
     this._loadLoaderDraft();
   }
 
-  _buildProfileDevicesStatus() {
-    if (!this._activeProfile) {
-      return "";
-    }
-
-    const stateMeta = this._profileDeviceStateMeta();
-    if (stateMeta.state === "active") {
-      const runningMeta =
-        (this._runningMetadata || {})[SCENARIO_PROFILE_EMISSIONS] || {};
-      const deviceCount =
-        runningMeta.devices?.length || stateMeta.profileCount || 0;
-      const label = deviceCount
-        ? `${deviceCount} devices emitting`
-        : stateMeta.description;
-      return `
-        <div style="margin-top:8px; padding:8px 12px; border-radius:8px; background:var(--primary-color, #3f51b5); color:var(--text-primary-color,#fff); display:flex; align-items:center; gap:8px; flex-wrap:wrap;">
-          <span>🔥 ${label}</span>
-          <button class="btn btn-secondary" data-action="stop-profile-devices">Stop all profile devices</button>
-        </div>`;
-    }
-
-    if (stateMeta.state === "reply") {
-      return `
-        <div style="margin-top:8px; padding:8px 12px; border-radius:8px; background:var(--info-color-light, rgba(2,136,209,0.15)); color:var(--primary-text-color); display:flex; align-items:center; gap:8px; flex-wrap:wrap;">
-          <span>🛰️ ${stateMeta.description}</span>
-        </div>`;
-    }
-
-    if (stateMeta.knownCount) {
-      return `
-        <div style="margin-top:8px; padding:8px 12px; border-radius:8px; background:var(--secondary-background-color); color:var(--secondary-text-color); display:flex; align-items:center; gap:8px; flex-wrap:wrap;">
-          <span>⏸️ ${stateMeta.description}</span>
-        </div>`;
-    }
-
-    return "";
-  }
-
   _getScenarioFieldMeta(scenarioId, key) {
     const schema = this._scenarioSchemas[scenarioId] || [];
     return schema.find((field) => field.key === key) || null;
@@ -244,94 +98,6 @@ class DeviceSimulatorCard extends RamsesBaseCard {
     }
 
     return [];
-  }
-
-  _renderDeviceZones(device) {
-    const zones = device?.zones || [];
-    if (!zones.length) {
-      return "";
-    }
-
-    const items = zones.map((zone) => {
-      const label = zone.label || `Zone ${zone.zone_id || "?"}`;
-      const controller = zone.controller ? `CTL ${zone.controller}` : null;
-      const roleLabel = (zone.roles || [])
-        .map((role) => (role === "devices" ? "device" : role))
-        .join(", ");
-      const badge = roleLabel ? `<span class="chip muted">${roleLabel}</span>` : "";
-      const sensor = zone.sensor ? `<span class="chip sensor">sensor ${zone.sensor}</span>` : "";
-
-      let members = "";
-      if ((zone.roles || []).includes("controller") && Array.isArray(zone.members)) {
-        const formattedMembers = zone.members
-          .map((member) => {
-            const memberRoles = member.roles?.includes("sensor")
-              ? "(sensor)"
-              : member.roles?.includes("actuator")
-                ? "(actuator)"
-                : "";
-            return `${member.id}${memberRoles ? ` ${memberRoles}` : ""}`;
-          })
-          .join(", ");
-        if (formattedMembers) {
-          members = `<div class="device-zone-members">Members: ${formattedMembers}</div>`;
-        }
-      }
-
-      return `
-        <div class="device-zone-row">
-          <div>
-            <strong>${label}</strong>${controller ? ` · ${controller}` : ""}
-          </div>
-          <div style="display:flex; gap:4px; flex-wrap:wrap; font-size:0.75em;">${badge}${sensor}</div>
-          ${members}
-        </div>`;
-    }).join("");
-
-    return `
-      <div class="device-zones" style="margin-top:12px;">
-        <div style="font-size:0.8em; color:var(--secondary-text-color); margin-bottom:4px;">Zone relationships</div>
-        <div class="device-zone-list" style="display:flex; flex-direction:column; gap:6px;">${items}</div>
-      </div>`;
-  }
-
-  // ========== REQUIRED BASE CLASS OVERRIDES ==========
-
-  getCardSize() {
-    return 6;
-  }
-
-  getFeatureName() {
-    return "device_simulator";
-  }
-
-  hasValidConfig() {
-    return true;
-  }
-
-  static getCardInfo() {
-    return {
-      type: "device-simulator-card",
-      name: "Device Simulator",
-      description: "Control and monitor the Ramses device simulator",
-      preview: true,
-      documentationURL: "https://github.com/wimpie70/ramses_extras",
-    };
-  }
-
-  static getStubConfig() {
-    return {
-      type: `custom:${this.getTagName()}`,
-      layout_options: {
-        grid_columns: 200,
-        rows: 1,
-      },
-    };
-  }
-
-  async _loadInitialState() {
-    await this._fetchData();
-    await this._fetchRfConfig();
   }
 
   _onConnected() {
@@ -1052,16 +818,16 @@ class DeviceSimulatorCard extends RamsesBaseCard {
         </div>
 
         <div class="content ${this._tab === "profiles" ? "active" : ""}" data-content="profiles">
-          ${this._buildProfiles()}
+          ${buildProfiles(this)}
         </div>
         <div class="content ${this._tab === "scenarios" ? "active" : ""}" data-content="scenarios">
-          ${this._buildScenarios()}
+          ${buildScenarios(this)}
         </div>
         <div class="content ${this._tab === "devices" ? "active" : ""}" data-content="devices">
-          ${this._buildDevices()}
+          ${buildDevices(this)}
         </div>
         <div class="content ${this._tab === "events" ? "active" : ""}" data-content="events">
-          ${this._buildEvents()}
+          ${buildEvents(this)}
         </div>
       </ha-card>
     `;
@@ -1553,201 +1319,6 @@ class DeviceSimulatorCard extends RamsesBaseCard {
 
   // ========== HTML BUILDERS ==========
 
-  _buildProfiles() {
-    const loaderCard = this._buildProfileLoaderCard();
-    const profileCards = this._profiles.length === 0
-      ? "<div>No profiles available</div>"
-      : this._profiles.map((p) => {
-          const reloadChecked = (this._profileReload[p.name] ?? true) ? " checked" : "";
-          const preloadChecked = (this._profilePreloadSchema[p.name] ?? true) ? " checked" : "";
-          const resetCacheChecked = (this._profileResetCache[p.name] ?? false) ? " checked" : "";
-          const eavesdropChecked = (this._profileEavesdrop[p.name] ?? false) ? " checked" : "";
-          const skipHydrateChecked = (this._profileSkipHydrate[p.name] ?? false) ? " checked" : "";
-          const clearLogChecked = (this._profileClearLog[p.name] ?? false) ? " checked" : "";
-          const autoAnswerChecked = (p.enable_auto_answer ?? true) ? " checked" : "";
-          const deleteButton = p.can_delete
-            ? `<button class="btn btn-secondary" data-action="delete-profile" data-profile="${p.name}" style="margin-left:auto;">Delete</button>`
-            : "";
-          const badges = `
-            <div style="display:flex; gap:4px; align-items:center; font-size:0.75em; color:var(--secondary-text-color);">
-              ${p.is_active ? "<span class=\"chip profile\">Active</span>" : ""}
-              ${p.is_builtin ? "<span class=\"chip source\">Built-in</span>" : ""}
-            </div>`;
-          return `
-            <div class="card ${this._activeProfile === p.name ? "active" : ""}">
-              <div style="display:flex; align-items:center; gap:8px; flex-wrap:wrap;">
-                <strong>${p.name}</strong>
-                ${badges}
-                ${deleteButton}
-              </div>
-              <div style="font-size: 0.85em; color: var(--secondary-text-color); margin-bottom: 8px;">${p.description || "No description"}</div>
-              <div style="display: flex; gap: 12px; align-items: center; flex-wrap: wrap; font-size:0.8em;">
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;">
-                  <input type="checkbox" data-action="reload-check" data-profile="${p.name}"${reloadChecked} />
-                  Reload RF
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;">
-                  <input type="checkbox" data-action="preload-check" data-profile="${p.name}"${preloadChecked} />
-                  Preload schema
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;">
-                  <input type="checkbox" data-action="reset-cache-check" data-profile="${p.name}"${resetCacheChecked} />
-                  Reset cache
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;" title="Enable ramses_rf eavesdrop so HVAC devices are promoted (FAN/REM/CO2/HUM) from observed traffic.">
-                  <input type="checkbox" data-action="eavesdrop-check" data-profile="${p.name}"${eavesdropChecked} />
-                  Enable eavesdrop
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;">
-                  <input type="checkbox" data-action="skip-hydrate-check" data-profile="${p.name}"${skipHydrateChecked} />
-                  Remove database file
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;">
-                  <input type="checkbox" data-action="clear-log-check" data-profile="${p.name}"${clearLogChecked} />
-                  Clear message log
-                </label>
-                <label style="display:flex; align-items:center; gap:4px; cursor:pointer;" title="Enable auto-answer when this profile is loaded.">
-                  <input type="checkbox" data-action="auto-answer-check" data-profile="${p.name}"${autoAnswerChecked} />
-                  Enable auto answer
-                </label>
-                <button class="btn btn-primary" data-action="load-profile" data-profile="${p.name}">Load</button>
-              </div>
-            </div>`;
-        }).join("");
-
-    const notice = this._profileNotice
-      ? `<div style="margin-top: 12px; padding: 8px 12px; background: var(--warning-color, #ff9800); color: white; border-radius: 6px; font-size: 0.85em; display: flex; justify-content: space-between; align-items: center; gap: 8px;">
-           <span>⚠️ ${this._profileNotice}</span>
-           <button data-action="dismiss-notice" style="background: none; border: none; color: white; cursor: pointer; font-size: 1.1em; line-height: 1; padding: 0 2px;">✕</button>
-         </div>`
-      : "";
-
-    return `${loaderCard}<div class="grid">${profileCards}</div>${notice}`;
-  }
-
-  _buildProfileLoaderCard() {
-    const conflicts = this._scenarioConflicts(SCENARIO_LOAD_PROFILE).filter((id) => ![SCENARIO_LOAD_PROFILE, SCENARIO_PROFILE_EMISSIONS].includes(id));
-    const conflictWarn = conflicts.length
-      ? `<div style="font-size: 0.75em; color: var(--warning-color, #ff9800); margin-top: 4px;">⚠️ Conflicts with: ${conflicts.join(", ")}</div>`
-      : "";
-    const params = this._scenarioParams[SCENARIO_LOAD_PROFILE] || {};
-    const nameField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "profile_name");
-    const yamlField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "profile_yaml");
-    const reloadField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "reload_ramses");
-    const preloadField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "preload_schema");
-    const resetField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "reset_rf_cache");
-    const eavesdropField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "enable_eavesdrop");
-    const skipHydrateField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "remove_database");
-    const clearLogField = this._getScenarioFieldMeta(SCENARIO_LOAD_PROFILE, "clear_message_log");
-
-    const profileName = (params.profile_name ?? this._activeProfile ?? nameField?.default ?? "").trim();
-    const yamlValue = params.profile_yaml ?? this._activeProfileYaml ?? "";
-    const hasYaml = Boolean((yamlValue || "").trim());
-    const reloadValue = params.reload_ramses ?? reloadField?.default ?? true;
-    const preloadValue = params.preload_schema ?? preloadField?.default ?? true;
-    const resetValue = params.reset_rf_cache ?? resetField?.default ?? false;
-    const eavesdropValue = params.enable_eavesdrop ?? eavesdropField?.default ?? false;
-    const skipHydrateValue = params.remove_database ?? skipHydrateField?.default ?? false;
-    const clearLogValue = params.clear_message_log ?? clearLogField?.default ?? false;
-
-    const profileNameInput = `
-      <div class="scenario-field">
-        <label for="loader-profile-name">${nameField?.label || "Profile name"}</label>
-        <input id="loader-profile-name" type="text" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="profile_name" value="${profileName}" placeholder="${nameField?.placeholder || "imported_profile"}" />
-      </div>`;
-
-    const yamlInput = `
-      <div class="scenario-field">
-        <label for="loader-profile-yaml">${yamlField?.label || "Profile YAML"} <span style="color:var(--error-color);">*</span></label>
-        <textarea id="loader-profile-yaml" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="profile_yaml" placeholder="${yamlField?.placeholder || "known_list:\n  32:150000:\n    class: FAN"}">${yamlValue}</textarea>
-      </div>`;
-
-    const reloadToggle = `
-      <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;">
-        <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="reload_ramses" data-type="checkbox" ${reloadValue ? "checked" : ""} />
-        ${reloadField?.label || "Reload RF"}
-      </label>`;
-
-    const preloadToggle = `
-      <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;">
-        <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="preload_schema" data-type="checkbox" ${preloadValue ? "checked" : ""} />
-        ${preloadField?.label || "Preload schema"}
-      </label>`;
-
-    const resetToggle = `
-      <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;">
-        <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="reset_rf_cache" data-type="checkbox" ${resetValue ? "checked" : ""} />
-        ${resetField?.label || "Reset RF cache"}
-      </label>`;
-
-    const eavesdropToggle = `
-      <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;" title="Enable ramses_rf eavesdrop so HVAC devices are promoted (FAN/REM/CO2/HUM) from observed traffic.">
-        <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="enable_eavesdrop" data-type="checkbox" ${eavesdropValue ? "checked" : ""} />
-        ${eavesdropField?.label || "Enable eavesdrop"}
-      </label>`;
-
-    const skipHydrateToggle = `
-      <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;">
-        <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="remove_database" data-type="checkbox" ${skipHydrateValue ? "checked" : ""} />
-        ${skipHydrateField?.label || "Remove database file"}
-      </label>`;
-
-    const clearLogToggle = clearLogField
-      ? `
-        <label style="font-size: 0.8em; display:flex; align-items:center; gap:4px; cursor:pointer;">
-          <input type="checkbox" data-action="scenario-param" data-scenario-id="${SCENARIO_LOAD_PROFILE}" data-field="clear_message_log" data-type="checkbox" ${clearLogValue ? "checked" : ""} />
-          ${clearLogField.label || "Clear message log"}
-        </label>
-        <div style="font-size:0.7em; color:var(--secondary-text-color); margin-top:-4px;">
-          ${clearLogField.description || "Clear the device simulator UI log before applying the profile."}
-        </div>`
-      : "";
-
-    return `
-      <div class="card" style="margin-bottom: 16px;" data-card="profile-loader">
-        <div style="display:flex; justify-content: space-between; align-items:center; gap:8px; flex-wrap:wrap;">
-          <strong>Load Ramses RF profile (YAML)</strong>
-          <span class="chip ${hasYaml ? "profile" : "muted"}" data-loader-chip>${hasYaml ? "Ready" : "Awaiting YAML"}</span>
-        </div>
-        <div style="font-size:0.85em; color: var(--secondary-text-color); margin-top:4px;">
-          Paste a known_devices YAML snippet to import a simulator profile and activate its devices.
-        </div>
-        ${this._activeProfile ? `<div style="font-size:0.8em; color: var(--secondary-text-color); margin-top:6px;">Active profile: <strong>${this._activeProfile}</strong>${this._activeProfileZones?.length ? ` · ${this._activeProfileZones.length} zones detected` : ""}</div>` : ""}
-        ${conflictWarn}
-        ${this._buildProfileDevicesStatus()}
-        ${profileNameInput}
-        ${yamlInput}
-        <div style="display:flex; flex-direction:column; gap:6px; font-size:0.8em;">
-          ${reloadToggle}
-          ${preloadToggle}
-          ${resetToggle}
-          ${eavesdropToggle}
-          ${skipHydrateToggle}
-          ${clearLogToggle}
-        </div>
-        <div style="margin-top: 8px; display:flex; gap:8px; flex-wrap:wrap;">
-          <button class="btn btn-primary" data-action="start-scenario" data-scenario-id="${SCENARIO_LOAD_PROFILE}" ${!hasYaml || conflicts.length ? "disabled" : ""}>Load</button>
-          <button class="btn btn-secondary" data-action="reset-loader" ${hasYaml ? "" : "disabled"}>Clear form</button>
-        </div>
-      </div>`;
-  }
-
-  _fmtHex(hex) {
-    if (!hex) return "";
-    return hex.match(/.{1,2}/g)?.join(" ") || hex;
-  }
-
-  _fmtDecoded(decoded) {
-    if (decoded == null) return null;
-    if (typeof decoded === "string") return decoded;
-    if (typeof decoded === "number" || typeof decoded === "boolean") return String(decoded);
-    try {
-      return JSON.stringify(decoded);
-    } catch {
-      return String(decoded);
-    }
-  }
-
   _mergeDeviceMessages(batch) {
     if (!Array.isArray(batch) || batch.length === 0) {
       return;
@@ -1827,415 +1398,6 @@ class DeviceSimulatorCard extends RamsesBaseCard {
       owned_by_profile: true,
       zones,
     };
-  }
-
-  _buildDeviceMsgPreview(deviceId) {
-    const messages = this._deviceMsgPreviews[deviceId];
-    if (!messages || !messages.length) {
-      return `<div class="msg-log-empty">No traffic yet — refresh to update</div>`;
-    }
-    const headerRow = `
-      <div class="msg-log-header-row">
-        <span>Recent traffic</span>
-        <button class="btn btn-secondary" data-action="clear-device-log" data-device-id="${deviceId}">Clear</button>
-      </div>`;
-    const header = `
-      <thead>
-        <tr>
-          <th>Time</th>
-          <th>Dir</th>
-          <th>Source</th>
-          <th>Verb</th>
-          <th>Code</th>
-          <th>Src</th>
-          <th>Dst</th>
-          <th>Bcast</th>
-          <th>Payload</th>
-        </tr>
-      </thead>`;
-    const originLabels = {
-      rf: "RF",
-      sim: "SIM",
-      auto_answer: "AUTO-RP",
-      auto_emit: "AUTO-I",
-    };
-    const rows = messages.slice().sort((a, b) => b.ts - a.ts).map((m) => {
-      const dir = m.direction === "inbound" ? "in" : "out";
-      const origin = m.origin || (dir === "in" ? "rf" : "sim");
-      const originLabel = originLabels[origin] || origin.toUpperCase();
-      const ts = new Date(m.ts * 1000).toLocaleTimeString([], { hour12: false, hour: "2-digit", minute: "2-digit", second: "2-digit", fractionalSecondDigits: 3 });
-      const hexPayload = this._fmtHex(m.payload);
-      const decoded = this._fmtDecoded(m.decoded_payload);
-      const payloadDisplay = decoded ?? hexPayload;
-      const payloadTitle = decoded ? `${decoded}\n\nraw: ${hexPayload}` : hexPayload;
-      const bcast = m.broadcast || "--:------";
-      return `<tr class="${dir} origin-${origin}">
-        <td class="msg-ts">${ts}</td>
-        <td class="msg-dir ${dir}">${dir === "in" ? "RX" : "TX"}</td>
-        <td><span class="msg-origin origin-${origin}" title="Frame origin: ${origin}">${originLabel}</span></td>
-        <td class="msg-verb">${m.verb}</td>
-        <td class="msg-code">${m.code}</td>
-        <td class="msg-addr" title="${m.src}">${m.src}</td>
-        <td class="msg-addr" title="${m.dst}">${m.dst}</td>
-        <td class="msg-addr" title="${bcast}">${bcast}</td>
-        <td class="msg-payload" title="${payloadTitle}">${payloadDisplay}</td>
-      </tr>`;
-    }).join("");
-    return `<div class="msg-log">${headerRow}<table class="msg-log-table">${header}<tbody>${rows}</tbody></table></div>`;
-  }
-
-  _buildDevices() {
-    const controls = `
-      <div class="grid device-controls">
-        ${this._buildManualInjectionCard()}
-        ${this._buildProfileEmissionsCard()}
-        ${this._buildResumeEmittersCard()}
-        ${this._buildPlaybackConversationCard()}
-        ${this._buildAutonomousSpeedCard()}
-      </div>`;
-
-    if (this._devices.length === 0) {
-      return `${controls}<div class="device-list-empty">No active devices. Use Manual Device Injection or start the profile emitters above to populate the simulator.</div>`;
-    }
-
-    const knownList = this._knownList();
-    const deviceCards = this._devices.map((d) => {
-      const isEmitting = d.emitting === true;
-      const isSilenced = d.suppress_autonomous;
-      const emitterLabel = isEmitting
-        ? "Emitter running"
-        : isSilenced
-          ? "Silenced"
-          : "Emitter idle";
-      const emitterChipClass = isEmitting
-        ? "chip emitting"
-        : isSilenced
-          ? "chip silenced"
-          : "chip idle";
-      const ownershipChip = d.owned_by_profile
-        ? `<span class="chip profile" title="Defined by the active profile">Profile</span>`
-        : `<span class="chip manual" title="Manually injected device">Manual</span>`;
-      const knownBadge = knownList[d.id]
-        ? `<span class="chip known" title="Device appears in profile known list">Known</span>`
-        : "";
-      const extraSource = d.source && !["manual", "profile"].includes(d.source)
-        ? `<span class="chip source" title="Origin scenario">${d.source}</span>`
-        : "";
-      const chipsMarkup = (d.excluded_codes || []).map((code) =>
-        `<span class="code-chip">${code}<button data-action="remove-code" data-device-id="${d.id}" data-code="${code}" title="Remove">✕</button></span>`
-      ).join("");
-      const checkedAttr = d.enabled ? " checked" : "";
-      const zoneMarkup = this._renderDeviceZones(d);
-      const emitterButton = isEmitting
-        ? `<button class="btn btn-secondary" data-action="silence-device" data-device-id="${d.id}">Stop emission</button>`
-        : `<button class="btn btn-secondary" data-action="resume-device" data-device-id="${d.id}">${isSilenced ? "Unsilence & resume" : "Resume emission"}</button>`;
-      return `
-        <div class="card">
-          <div style="display: flex; justify-content: space-between; align-items: center; gap: 8px; flex-wrap: wrap;">
-            <div><strong>${d.id}</strong> <span style="color: var(--secondary-text-color); font-size: 0.85em;">${d.type}</span></div>
-            <div style="display:flex; gap:4px; flex-wrap: wrap;">${ownershipChip}${knownBadge}${extraSource}</div>
-          </div>
-          <div class="device-actions">
-            <label class="toggle">
-              <ha-switch data-action="toggle-device" data-device-id="${d.id}"${checkedAttr}></ha-switch>
-              <span>${d.enabled ? "Enabled" : "Disabled"}</span>
-            </label>
-          </div>
-          <div class="device-emitter" title="Autonomous emitter state">
-            <span class="${emitterChipClass}">${emitterLabel}</span>
-            ${emitterButton}
-            <button class="btn btn-secondary" data-action="discover-device" data-device-id="${d.id}">Discover capabilities</button>
-          </div>
-          <div style="font-size: 0.8em; color: var(--secondary-text-color); margin-top: 8px;">Excluded codes:</div>
-          <div class="codes-row">${chipsMarkup || `<span class="chip muted">none</span>`}</div>
-          <div class="add-code">
-            <input type="text" maxlength="4" placeholder="1FC9"
-              data-action="code-input" data-device-id="${d.id}"
-              value="${this._newCodeInput[d.id] || ""}" />
-            <button class="btn btn-primary" data-action="add-code" data-device-id="${d.id}">+ Exclude</button>
-          </div>
-          ${zoneMarkup}
-          <div style="margin-top:10px;">
-            <div style="font-size:0.75em; color:var(--secondary-text-color); margin-bottom:3px; font-weight:600;">Recent traffic</div>
-            ${this._buildDeviceMsgPreview(d.id)}
-          </div>
-        </div>`;
-    }).join("");
-
-    return `${controls}<div class="grid devices-grid">${deviceCards}</div>`;
-  }
-
-  _buildManualInjectionCard() {
-    const manualCount = this._manualDeviceCount();
-    const conflicts = this._scenarioConflicts(SCENARIO_MANUAL_DEVICE);
-    const conflictWarn = conflicts.length
-      ? `<div style="font-size: 0.75em; color: var(--warning-color, #ff9800); margin-top: 4px;">⚠️ Conflicts with: ${conflicts.join(", ")}</div>`
-      : "";
-    const statusChip = manualCount
-      ? `<span class="chip manual">${manualCount} active</span>`
-      : `<span class="chip muted">Idle</span>`;
-    const knownList = this._knownList();
-    const templateOptions = Object.entries(knownList).map(([deviceId, meta]) => {
-      const slug = (meta?.class || "FAN").toUpperCase();
-      const variant = meta?.variant || meta?.variant_id || "default";
-      return `<option value="${deviceId}" data-slug="${slug}" data-variant="${variant}">${deviceId} • ${slug}</option>`;
-    }).join("");
-    const templateSelect = templateOptions
-      ? `<label style="font-size: 0.8em; display:flex; flex-direction:column; gap:4px;">
-            <span>Preset from profile</span>
-            <select data-action="manual-template" style="font-size:0.85em; padding:4px; border:1px solid var(--divider-color); border-radius:4px; background:var(--card-background-color); color:var(--primary-text-color);">
-              <option value="">Select device...</option>
-              ${templateOptions}
-            </select>
-         </label>`
-      : "";
-
-    return `
-      <div class="card">
-        <div style="display:flex; justify-content: space-between; align-items:center; flex-wrap: wrap; gap:8px;">
-          <strong>Manual Device Injection</strong>
-          ${statusChip}
-        </div>
-        <div style="font-size:0.85em; color:var(--secondary-text-color); margin-top:4px;">Inject an ad-hoc device that emits its periodic frames. Use profile presets to auto-fill slug/device id.</div>
-        ${templateSelect}
-        ${this._buildScenarioParamsById(SCENARIO_MANUAL_DEVICE)}
-        ${conflictWarn}
-        <div style="margin-top:8px; display:flex; gap:8px; flex-wrap:wrap;">
-          <button class="btn btn-primary" data-action="start-scenario" data-scenario-id="${SCENARIO_MANUAL_DEVICE}" ${conflicts.length ? "disabled" : ""}>Inject Device</button>
-          <button class="btn btn-secondary" data-action="stop-scenario" data-scenario-id="${SCENARIO_MANUAL_DEVICE}" ${manualCount ? "" : "disabled"}>Stop manual devices</button>
-        </div>
-      </div>`;
-  }
-
-  _buildProfileEmissionsCard() {
-    const running = (this._runningScenarios || []).includes(SCENARIO_PROFILE_EMISSIONS);
-    const conflicts = this._scenarioConflicts(SCENARIO_PROFILE_EMISSIONS);
-    const conflictWarn = conflicts.length
-      ? `<div style="font-size: 0.75em; color: var(--warning-color, #ff9800); margin-top: 4px;">⚠️ Conflicts with: ${conflicts.join(", ")}</div>`
-      : "";
-    const knownList = this._knownList();
-    const summary = this._profileDeviceSummary || [];
-    const counts = this._profileDeviceCounts || null;
-    const knownCount = counts?.known ?? (summary.length || Object.keys(knownList).length);
-    const profileCount = counts?.active ?? this._profileDeviceCount();
-    const stateMeta = this._profileDeviceStateMeta(profileCount, knownCount);
-    const statusChip = `<span class="${stateMeta.chipClass}">${stateMeta.label}</span>`;
-    const disableStart = !this._activeProfile || !knownCount || conflicts.length;
-    const startDisabled = disableStart || running;
-    const stopDisabled = profileCount === 0;
-    const summaryText = stateMeta.description;
-
-    const startButton = `<button class="btn btn-primary" data-action="start-scenario" data-scenario-id="${SCENARIO_PROFILE_EMISSIONS}" ${startDisabled ? "disabled" : ""}>Start profile emitters</button>`;
-    const stopButton = `<button class="btn btn-secondary" data-action="stop-profile-devices" ${stopDisabled ? "disabled" : ""}>Stop profile emitters</button>`;
-    const buttonRow = `${startButton}${stopButton}`;
-
-    const missingDevices = summary.filter((entry) => !entry.active);
-    const missingMarkup = missingDevices.length
-      ? `
-        <div class="profile-missing">
-          <strong>${missingDevices.length === 1 ? "1 device" : `${missingDevices.length} devices`} not emitting</strong>
-          ${missingDevices
-            .map(
-              (entry) => `
-                <div class="profile-missing-entry">
-                  <span>${entry.id}${entry.class ? ` • ${entry.class}` : ""}</span>
-                  <button class="btn btn-secondary" data-action="activate-profile-device" data-device-id="${entry.id}">
-                    Activate
-                  </button>
-                </div>`,
-            )
-            .join("")}
-        </div>`
-      : "";
-
-    return `
-      <div class="card">
-        <div style="display:flex; justify-content: space-between; align-items:center; flex-wrap:wrap; gap:8px;">
-          <strong>Profile emitters</strong>
-          ${statusChip}
-        </div>
-        <div style="font-size:0.85em; color:var(--secondary-text-color); margin-top:4px;">${summaryText} These devices emit their periodic "random" frames directly from the active profile.</div>
-        ${conflictWarn}
-        <div style="margin-top:8px; display:flex; gap:8px; flex-wrap:wrap;">${buttonRow}</div>
-        ${missingMarkup}
-      </div>`;
-  }
-
-  _buildResumeEmittersCard() {
-    const devices = this._devices || [];
-    // Resume should work on any active device whose emitter is idle, including
-    // devices that arrived disabled via playback/discovery or were silenced.
-    // The backend `async_resume_device` re-enables and unsilences the device.
-    const resumeCandidates = devices.filter((d) => !d.emitting);
-    const silencedCount = devices.filter((d) => d.suppress_autonomous).length;
-    const emittingCount = devices.filter((d) => d.emitting).length;
-    const chipClass = resumeCandidates.length
-      ? "chip reply"
-      : devices.length
-        ? "chip emitting"
-        : "chip muted";
-    const chipLabel = resumeCandidates.length
-      ? `${resumeCandidates.length} idle`
-      : devices.length
-        ? "All emitting"
-        : "No active devices";
-    const desc = resumeCandidates.length
-      ? "Start autonomous emissions for the devices you just activated via playback or discovery."
-      : devices.length
-        ? "Every active device already has its emitter running."
-        : "Activate devices first (via playback, manual injection, or profiles) to resume their emitters.";
-    const sampleList = resumeCandidates.slice(0, 4)
-      .map((d) => `<span class="chip muted">${d.id}</span>`)
-      .join("");
-    const extraCount = resumeCandidates.length > 4
-      ? `<span class="chip muted">+${resumeCandidates.length - 4} more</span>`
-      : "";
-    const silencedNote = silencedCount
-      ? `<div class="profile-missing" style="margin-top:8px;">
-          <strong>${silencedCount} device${silencedCount === 1 ? " is" : "s are"} silenced</strong>
-          <span>Use the per-device "Unsilence & resume" button to re-enable them.</span>
-        </div>`
-      : "";
-    return `
-      <div class="card">
-        <div style="display:flex; justify-content: space-between; align-items:center; flex-wrap:wrap; gap:8px;">
-          <strong>Resume emitters for active devices</strong>
-          <span class="${chipClass}">${chipLabel}</span>
-        </div>
-        <div style="font-size:0.85em; color:var(--secondary-text-color); margin-top:4px;">${desc}</div>
-        <div style="margin-top:8px; display:flex; gap:8px; flex-wrap:wrap;">
-          <button class="btn btn-secondary" data-action="resume-all-devices" ${resumeCandidates.length ? "" : "disabled"}>Start emitters for active devices</button>
-          <button class="btn btn-secondary" data-action="silence-all-devices" ${emittingCount ? "" : "disabled"}>Stop emitters for active devices</button>
-        </div>
-        ${resumeCandidates.length ? `<div style="margin-top:8px; display:flex; gap:6px; flex-wrap:wrap;">${sampleList}${extraCount}</div>` : ""}
-        ${silencedNote}
-      </div>`;
-  }
-
-  _formatAutonomousSpeedLabel(speed) {
-    const value = Number(speed) || 1;
-    if (Math.abs(value - 1) < 0.01) {
-      return "1× normal";
-    }
-    if (value > 1) {
-      const magnitude = value >= 10 ? value.toFixed(0) : value.toFixed(1);
-      return `${magnitude}× faster`;
-    }
-    const slower = 1 / value;
-    const text = slower >= 10 ? slower.toFixed(0) : slower.toFixed(1);
-    return `${text}× slower`;
-  }
-
-  _buildPlaybackConversationCard() {
-    const playbacks = this._savedPlaybacks || [];
-    const runMeta = (this._runningMetadata || {})[SCENARIO_DEVICE_PLAYBACK] || null;
-    const isRunning = Boolean(runMeta);
-    const isPaused = Boolean(runMeta && runMeta.paused);
-    const selection = this._playbackSelection || (playbacks[0] && playbacks[0].id) || "";
-    const params = this._scenarioParams?.[SCENARIO_DEVICE_PLAYBACK] || {};
-    const loopsValue = params.loops ?? 1;
-
-    const optionList = playbacks.length
-      ? playbacks.map((p) => `<option value="${p.id}"${p.id === selection ? " selected" : ""}>${p.id} (${p.frames} frames)</option>`).join("")
-      : `<option value="" disabled selected>No saved playbacks</option>`;
-
-    const statusBadge = isRunning
-      ? `<span class="chip" style="background:${isPaused ? "var(--warning-color,#ff9800)" : "var(--success-color,#4caf50)"}; color:white;">
-          ${isPaused ? "Paused" : "Playing"}${runMeta?.conversation ? ` · ${runMeta.conversation}` : ""}
-        </span>`
-      : `<span class="chip muted" title="No playback is running">Idle</span>`;
-
-    const playDisabled = !selection || isRunning ? " disabled" : "";
-    const pauseDisabled = !isRunning || isPaused ? " disabled" : "";
-    const resumeDisabled = !isRunning || !isPaused ? " disabled" : "";
-    const stopDisabled = !isRunning ? " disabled" : "";
-
-    return `
-      <div class="card">
-        <div style="display:flex; justify-content:space-between; align-items:center; gap:8px;">
-          <strong>Playback conversation</strong>
-          ${statusBadge}
-        </div>
-        <div style="font-size:0.85em; color:var(--secondary-text-color); margin-top:4px;">
-          Replay a saved conversation. Speed follows the slider below (live).
-        </div>
-        <div style="display:flex; flex-direction:column; gap:6px; margin-top:8px;">
-          <label style="display:flex; flex-direction:column; gap:2px; font-size:0.75em;">
-            <span>Conversation</span>
-            <select data-action="playback-select">${optionList}</select>
-          </label>
-          <label style="display:flex; flex-direction:column; gap:2px; font-size:0.75em;">
-            <span>Loops</span>
-            <input type="number" min="1" step="1" value="${loopsValue}"
-                   data-action="scenario-param"
-                   data-scenario-id="${SCENARIO_DEVICE_PLAYBACK}"
-                   data-field="loops"
-                   data-type="number" />
-          </label>
-          <label style="display:flex; flex-direction:column; gap:2px; font-size:0.75em;">
-            <span>Fixed gap between frames (s)</span>
-            <input type="number" min="0" step="0.01"
-                   value="${this._playbackInterMsgDelay ?? ""}"
-                   placeholder="Leave blank to use recorded timing"
-                   data-action="playback-inter-msg-delay" />
-          </label>
-          <label style="display:flex; align-items:center; gap:6px; font-size:0.75em; cursor:pointer;"
-                 title="Skip recorded RP frames so Auto Answer (if enabled) replies instead.">
-            <input type="checkbox"
-                   data-action="playback-skip-answers"
-                   ${this._playbackSkipAnswers ? "checked" : ""} />
-            <span>Skip recorded answers (let Auto Answer reply)</span>
-          </label>
-        </div>
-        <div style="display:flex; gap:6px; flex-wrap:wrap; margin-top:8px;">
-          <button class="btn btn-primary" data-action="start-saved-playback" data-identifier="${selection}"${playDisabled}>▶ Play</button>
-          <button class="btn btn-secondary" data-action="pause-scenario" data-scenario-id="${SCENARIO_DEVICE_PLAYBACK}"${pauseDisabled}>⏸ Pause</button>
-          <button class="btn btn-secondary" data-action="resume-scenario" data-scenario-id="${SCENARIO_DEVICE_PLAYBACK}"${resumeDisabled}>▶ Resume</button>
-          <button class="btn btn-danger" data-action="stop-scenario" data-scenario-id="${SCENARIO_DEVICE_PLAYBACK}"${stopDisabled}>⏹ Stop</button>
-        </div>
-      </div>`;
-  }
-
-  _buildAutonomousSpeedCard() {
-    const current = Number(this._autonomousSpeed) || 1;
-    const pending = Number.isFinite(this._pendingSpeed) ? this._pendingSpeed : current;
-    const sliderValue = Math.min(1, Math.max(0.01, pending || 1));
-    const presetValues = [2, 1, 0.5, 0.25, 0.1, 0.05, 0.02];
-
-    const badge = this._speedSaving
-      ? '<span class="chip muted">Saving…</span>'
-      : `<span class="chip profile">${this._formatAutonomousSpeedLabel(current)}</span>`;
-
-    const presetButtons = presetValues
-      .map((value) => {
-        const active = Math.abs(current - value) < 0.001 ? "btn-primary" : "btn-secondary";
-        return `<button class="btn ${active}" data-action="speed-preset" data-speed="${value}">${this._formatAutonomousSpeedLabel(value)}</button>`;
-      })
-      .join("");
-
-    return `
-      <div class="card speed-card">
-        <div style="display:flex; justify-content:space-between; align-items:center; flex-wrap:wrap; gap:8px;">
-          <strong>Autonomous Emission Speed</strong>
-          ${badge}
-        </div>
-        <div style="font-size:0.85em; color:var(--secondary-text-color);">
-          Lower than 1× slows emitters down, higher than 1× speeds them up. The slider covers 0.01–1×; use the number field or presets for faster modes.
-        </div>
-        <div class="speed-controls">
-          <label style="display:flex; flex-direction:column; gap:4px; font-size:0.75em;">
-            <span>Slowdown (0.01–1×)</span>
-            <input type="range" min="0.01" max="1" step="0.01" value="${sliderValue}" data-action="speed-slider" />
-          </label>
-          <label style="display:flex; flex-direction:column; gap:4px; font-size:0.75em;">
-            <span>Exact multiplier</span>
-            <input type="number" min="0.01" max="100" step="0.01" value="${pending}" data-action="speed-input" />
-          </label>
-        </div>
-        <div style="display:flex; flex-wrap:wrap; gap:6px;">
-          ${presetButtons}
-        </div>
-      </div>`;
   }
 
   _handleScenarioParamInput(input) {
@@ -2425,228 +1587,6 @@ class DeviceSimulatorCard extends RamsesBaseCard {
       // eslint-disable-next-line no-console
       console.warn("Device Simulator: failed to persist loader draft", error);
     }
-  }
-
-  _buildScenarios() {
-    const registry = this._scenarioRegistry;
-    const ids = Object.keys(registry);
-    if (!ids.length) return `<div style="color: var(--secondary-text-color); padding: 8px 0;">No scenarios available.</div>`;
-
-    const psChecked = this._preserveState ? " checked" : "";
-    const psCard = `
-      <div class="card ${this._preserveState ? "active" : ""}">
-        <div style="display: flex; justify-content: space-between; align-items: center;">
-          <div><strong>Preserve state on reload / Clean restart</strong></div>
-          <label class="toggle">
-            <ha-switch data-action="toggle-preserve-state"${psChecked}></ha-switch>
-          </label>
-        </div>
-        <div style="font-size: 0.85em; color: var(--secondary-text-color); margin-top: 4px;">When on: simulator state (auto-answer, answer unknown devices) persists across reloads. When off (Clean restart): state is reset on reload (auto-answer disabled, clean known list).</div>
-      </div>`;
-
-    const aaChecked = this._autoAnswer ? " checked" : "";
-    const aaCard = `
-      <div class="card ${this._autoAnswer ? "active" : ""}">
-        <div style="display: flex; justify-content: space-between; align-items: center;">
-          <div><strong>Auto Answer (RQ&#8594;RP)</strong></div>
-          <label class="toggle">
-            <ha-switch data-action="toggle-auto-answer"${aaChecked}></ha-switch>
-          </label>
-        </div>
-        <div style="font-size: 0.85em; color: var(--secondary-text-color); margin-top: 4px;">When off: simulator receives RQ frames but never replies &#8212; simulates broken ESP or powered-off device.</div>
-      </div>`;
-
-    const auChecked = this._answerUnknownDevices ? " checked" : "";
-    const auCard = `
-      <div class="card ${this._answerUnknownDevices ? "active" : ""}">
-        <div style="display: flex; justify-content: space-between; align-items: center;">
-          <div><strong>Answer Unknown Devices</strong></div>
-          <label class="toggle">
-            <ha-switch data-action="toggle-answer-unknown"${auChecked}></ha-switch>
-          </label>
-        </div>
-        <div style="font-size: 0.85em; color: var(--secondary-text-color); margin-top: 4px;">When on: simulator responds to RQ frames for devices not in the active device list. Uses device DB to infer device type and generate responses.</div>
-      </div>`;
-
-    const rfStatus = this._rfEnforceKnownList
-      ? `<span style="color: var(--warning-color); font-weight: 600;">ENFORCED</span>`
-      : `<span style="color: var(--success-color); font-weight: 600;">NOT ENFORCED</span>`;
-    const rfCard = `
-      <div class="card" style="background: var(--secondary-background-color);">
-        <div style="display: flex; justify-content: space-between; align-items: center;">
-          <div><strong>RF Client: Accept only known devices</strong></div>
-          <div>${rfStatus}</div>
-        </div>
-        <div style="font-size: 0.85em; color: var(--secondary-text-color); margin-top: 4px;">RF client ${this._rfEnforceKnownList ? "will only accept messages from devices in its known_list" : "accepts messages from any device"}. ${this._rfKnownListEnabled ? `Known list is configured with entries.` : "No known list configured."}</div>
-      </div>`;
-
-    const scenarioCards = ids
-      .filter(id =>
-        id !== "auto_answer"
-        && id !== SCENARIO_MANUAL_DEVICE
-        && id !== SCENARIO_LOAD_PROFILE
-        && id !== SCENARIO_PROFILE_EMISSIONS
-        && id !== "device_suite"
-        && id !== SCENARIO_DEVICE_PLAYBACK
-      )
-      .map((id) => {
-        const meta = registry[id];
-        const isRunning = (this._runningScenarios || []).includes(id);
-        const runningMeta = (this._runningMetadata || {})[id] || {};
-        const isPaused = Boolean(runningMeta.paused);
-        const conflicts = this._scenarioConflicts(id);
-        const conflictWarn = conflicts.length
-          ? `<div style="font-size: 0.75em; color: var(--warning-color, #ff9800); margin-top: 4px;">&#9888; Conflicts with: ${conflicts.join(", ")}</div>`
-          : "";
-        const disabledAttr = conflicts.length ? " disabled" : "";
-        let actionBtns;
-        if (meta.toggleable) {
-          // Toggleable scenarios render as a switch (like auto_answer), no Run button.
-          const checked = isRunning ? " checked" : "";
-          const disSwitch = (!isRunning && conflicts.length) ? " disabled" : "";
-          actionBtns = `
-            <label class="toggle" style="display:flex; align-items:center; gap:6px;">
-              <ha-switch data-action="toggle-scenario" data-scenario-id="${id}"${checked}${disSwitch}></ha-switch>
-              <span style="font-size:0.75em; color:var(--secondary-text-color);">${isRunning ? "enabled" : "disabled"}</span>
-            </label>`;
-        } else if (isRunning) {
-          actionBtns = `<button class="btn btn-danger" data-action="stop-scenario" data-scenario-id="${id}">Stop</button>`;
-        } else {
-          actionBtns = `<button class="btn btn-primary" data-action="start-scenario" data-scenario-id="${id}"${disabledAttr}>Run</button>`;
-        }
-        const runningBadge = isRunning
-          ? `<span style="font-size:0.75em; padding:2px 8px; border-radius:10px; background:${isPaused ? "var(--warning-color,#ff9800)" : "var(--success-color,#4caf50)"}; color:white;">${isPaused ? "paused" : "active"}</span>`
-          : "";
-        return `
-          <div class="card ${isRunning ? "active" : ""}">
-            <div style="display: flex; justify-content: space-between; align-items: center;">
-              <strong>${meta.label || id}</strong>
-              ${runningBadge}
-            </div>
-            <div class="scenario-description">${meta.description || "No description"}</div>
-            ${conflictWarn}
-            ${this._buildScenarioParamsById(id)}
-            <div style="margin-top: 8px; display: flex; gap: 8px; flex-wrap: wrap; align-items: center;">${actionBtns}</div>
-          </div>`;
-      }).join("");
-
-    const playbackCard = this._buildConversationPlaybackSettingsCard();
-    return `<div class="grid">${psCard}${aaCard}${auCard}${rfCard}${playbackCard}${scenarioCards}</div>`;
-  }
-
-  _buildSavedPlaybacksPanel() {
-    const list = this._savedPlaybacks || [];
-    const query = (this._playbackSearchQuery || "").toLowerCase().trim();
-    const filteredList = query
-      ? list.filter((p) => p.id.toLowerCase().includes(query))
-      : list;
-    const builtinCount = list.filter((p) => p.builtin).length;
-    const userCount = list.length - builtinCount;
-    const header = `
-      <div style="margin-top:8px; font-size:0.85em; color:var(--secondary-text-color);">
-        <strong>Saved playbacks</strong>
-        <span style="margin-left:6px;">(${builtinCount} built-in, ${userCount} imported)</span>
-      </div>
-      <div style="margin-top:4px;">
-        <input type="text"
-               placeholder="Search conversations..."
-               value="${this._playbackSearchQuery || ""}"
-               data-action="playback-search"
-               style="width:100%; padding:4px 8px; border:1px solid var(--divider-color,#eee); border-radius:4px; background:var(--card-background-color); color:var(--primary-text-color); font-size:0.8em;" />
-      </div>`;
-    if (!list.length) {
-      return `${header}
-        <div style="font-size:0.8em; color:var(--secondary-text-color); padding:4px 0;">
-          None yet. Paste a log above and click <em>Save</em> to keep it.
-        </div>`;
-    }
-    if (!filteredList.length) {
-      return `${header}
-        <div style="font-size:0.8em; color:var(--secondary-text-color); padding:4px 0;">
-          No matching conversations.
-        </div>`;
-    }
-    const rows = filteredList.map((p) => {
-      const builtinTag = p.builtin
-        ? `<span class="chip muted" style="font-size:0.65em;">built-in</span>`
-        : "";
-      const editBtn = `<button class="btn btn-secondary" style="padding:2px 8px; font-size:0.75em;"
-               data-action="load-playback-editor" data-identifier="${p.id}"
-               title="Load frames into the editor above (${p.builtin ? "saves as a new copy" : "edit and save as new name"})">Edit</button>`;
-      const deleteBtn = p.builtin
-        ? ""
-        : `<button class="btn btn-danger" style="padding:2px 8px; font-size:0.75em;"
-                   data-action="delete-saved-playback" data-identifier="${p.id}">Delete</button>`;
-      return `
-        <div style="display:flex; justify-content:space-between; align-items:center; padding:4px 0; border-top:1px solid var(--divider-color,#eee); gap:6px;">
-          <div style="display:flex; flex-direction:column; font-size:0.8em; flex:1;">
-            <div style="display:flex; align-items:center; gap:6px;">
-              <strong>${p.id}</strong>${builtinTag}
-            </div>
-            <span style="color:var(--secondary-text-color);">
-              ${p.frames} frame${p.frames === 1 ? "" : "s"} · ${(p.peers || []).join(", ") || "no peers"}
-            </span>
-          </div>
-          <div style="display:flex; gap:4px;">${editBtn}${deleteBtn}</div>
-        </div>`;
-    }).join("");
-    return `${header}<div style="margin-top:4px; max-height:300px; overflow-y:auto;">${rows}</div>`;
-  }
-
-  _buildConversationPlaybackSettingsCard() {
-    const draft = this._playbackImportDraft || {};
-    const logContent = draft.log_content || "";
-    const nameValue = draft.name || "";
-    const canSave = Boolean(logContent.trim());
-    const savingState = this._playbackSaving
-      ? `<span class="chip muted" style="margin-left:6px;">Saving…</span>`
-      : "";
-    return `
-      <div class="card">
-        <div style="display:flex; justify-content:space-between; align-items:center;">
-          <strong>Conversation Playback</strong>
-          ${savingState}
-        </div>
-        <div class="scenario-description">
-          Import a ramses.log and save it as a reusable conversation. Run/pause/stop playback from the <em>Devices</em> tab.
-        </div>
-        <div class="scenario-field">
-          <label for="playback-log-content">Log content (paste ramses.log)</label>
-          <textarea id="playback-log-content"
-                    data-action="playback-import-field" data-field="log_content"
-                    rows="6" placeholder="Paste ramses.log content here...">${logContent}</textarea>
-        </div>
-        <div class="scenario-field">
-          <label for="playback-name">Save as name</label>
-          <input id="playback-name" type="text"
-                 data-action="playback-import-field" data-field="name"
-                 placeholder="my_playback" value="${nameValue}" />
-        </div>
-        ${this._buildSavedPlaybacksPanel()}
-        <div style="margin-top:8px; display:flex; gap:8px; flex-wrap:wrap;">
-          <button class="btn btn-primary" data-action="save-playback-import" ${canSave ? "" : "disabled"}>Save</button>
-          <button class="btn btn-secondary" data-action="clear-playback-import" ${canSave || nameValue ? "" : "disabled"}>Clear</button>
-        </div>
-      </div>`;
-  }
-
-  _buildEvents() {
-    const events = this._events.length === 0
-      ? `<div style="padding: 16px; text-align: center; color: var(--secondary-text-color);">No events</div>`
-      : this._events.map((e) =>
-          `<div class="event ${e.type === "unavailable" ? "unavailable" : ""}">
-             <span class="event-type ${e.type}">${e.type}</span>
-             <span>${e.message}</span>
-           </div>`
-        ).join("");
-    return `
-      <div class="stats">
-        <div class="stat"><div class="stat-value">${this._stats.rx}</div><div>RX</div></div>
-        <div class="stat"><div class="stat-value">${this._stats.tx}</div><div>TX</div></div>
-        <div class="stat"><div class="stat-value">${this._stats.devices}</div><div>Devices</div></div>
-        <div class="stat"><div class="stat-value">${this._stats.active}</div><div>Active</div></div>
-      </div>
-      <div class="event-log">${events}</div>`;
   }
 }
