+++ docs/notepad.txt
@@ -94,15 +94,19 @@ Shouldn't we filter this ? we know from our database what belongs to what
 ### when a fan replies to a REM, then it must be bound....
 -> we should mark it bound, maybe even add it to _rf as bound ?
 
---- next to do:
-
 ### scenario: rotate log files: we get reports of crashes or losing devices.
 we should be able to simulate and test this
 
 
+--- next to do:
+
 - refactor device-simulator-card in smaller files:
   - tab files:
     - with sub-components
+  keep run on devices, rest to scenarios
+  remove seperate answer unknown devices (controlled by profiles)
+  move preserve state to profiles
+
 
 -- other:
